Support for ignoring extraneous fields in rest request body (#1608)
## Why make this change?
Add support for allowing extraneous fields in request body as per:
https://github.com/Azure/data-api-builder/issues/1606.

## What is this change?
DAB will expose another CLI option in the `init` command
`rest.request-body-strict`, which when set to true/false, would set the
property `runtime.rest.request-body-strict` as `true`/`false` in the
generated runtime config file. What this essentially means is that, as
part of the REST mutations (POST/PUT/PATCH), there can be extraneous
fields present in the request body and they will be ignored. Default
behavior of DAB will continue to be the same as what it is today, i.e.
any extraneous fields in the request body would return an error. So to
say, the default value of `runtime.rest.request-body-strict` is `true`.

## Additional change
Refactored `RequestValidator.cs` class to be a singleton dependent on
`ISqlMetaDataProvider` and `RuntimeConfigProvider` services. Better
presents its dependency on the two services and simplifies the code.

## What are the different extraneous fields?
1. Fields in request body that do not map to any backing column in the
table/view. They will be ignored.
2. Fields that are repeated - present as PK in the request URL for
PUT/PATCH requests and also present in the request body. The ones in the
request body will be ignored.
3. Read-only fields. The support for this will be added in a subsequent
PR: https://github.com/Azure/data-api-builder/pull/1596.
-> Read-only field are: **Computed** (or Generated) / **AutoGenerated**
fields in MsSql/MySql/PgSql
 -> **timestamp** fields in MsSql

## Questions:

**1. What about authorization of extraneous fields?**
When operating in flexible mode, the user is _not_ authorized for
extraneous fields NOT defined on table because those fields are ignored.
For all other cases, authorization goes as usual.

## How was this tested?

- [x] Integration Tests - Added tests to validate that extraneous fields
are allowed in the request body for all the 3 relational DBs in the
`{*}RestBodyNonStrictModeTests.cs` classes inheriting from the base
class `RestBodyNonStrictModeTests.cs`. (* = MsSql,MySql,PostgreSql)
- [x] Unit Tests - Added unit test to validate the functionality of the
newly added --`rest.request-body-strict` feature in the init command.

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>