[Cherrypick] Fix setting cosmos configuration at runtime by checking whether the graphql schema is set (#1746)
## Context
Cherry-picking #1739 from `main` to `release/0.9` which includes the
merge conflicts resolved for merging into main. Remainder of description
copied from #1739 PR description.

## Why make this change?
Failing to set the configuration at runtime if there is not "schema"
file set in the config. This is not required when setting the config at
runtime since the graphql schema is set already.

## How was this tested?

- [x] Validated locally
- [x] Updated the tests so the schema property is removed from the
config. Verified that the tests were failing before the fix and passing
after.

Co-authored-by: Mathieu Tremblay <mathieu.tremblay1994@gmail.com>
Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>