Unhandled Exception for malformed connection string bug fix (#1741)
## Why make this change?

Closes https://github.com/Azure/data-api-builder/issues/1736

When a connection string was not properly formed, in the case of `MSSQL`
database types it was possible to have an exception when modifying the
connection string as part of our normal code. However, that exception
was not being handled and would cause an ungraceful crash of the
application. We now handle this exception.

## What is this change?

Include `DataApiBuilderException` in the exception handling for where we
modify the connection string in this particular case. And include a
regression test case.

New error messaging can be seen here:


![image](https://github.com/Azure/data-api-builder/assets/93220300/887cb26e-dd88-4c48-b577-8b242ad06c82)


It should also be noted that some tests had expected behavior that did
not account for our new exception handling. Some test cases have been
changed to `postgresql` data base type to avoid conflicts with our new
post processing error handling, when that new error handling is not
relevant to what those tests were covering. Since the post processing
only happens in the following conditional, this is an easy way to keep
our test coverage without major changes.


![image](https://github.com/Azure/data-api-builder/assets/93220300/57d6cb73-8398-4a66-889a-a6c141677c75)



Our tests for error checking bad connection strings for mssql was also
relying on having the actual exception available, but we now have
exception handling that will catch that and return a different exception
than the test is expecting. To keep this test working we redirect
standard error and then checking against that messaging for those cases
where we are both mssql db type and have a malformed connection string .
The remaining difference for those cases that do not have valid
connection strings and are mssql db type is that we have an
`InternalServerError` as the status code now, so we modify the test to
look for that instead in this case.

## How was this tested?

A regression test is added to the unit tests for the runtime config
loader. This is where the change happened, and the new case covers a
malformed connection string, verifying that we handle the exception
correctly by validating that we return false from the function that was
throwing the unhandled exception.

## Sample Request(s)


You can manually repro with the instructions from the linked issue this
closes which are the following steps:

1. After building the solution, from the root of the repo: cd
src\out\cli\net6.0
2. Microsoft.DataApiBuilder.exe init --database-type mssql
--connection-string "abcd"
3. Microsoft.DataApiBuilder.exe add Book --source dbo.books
--permissions "anonymous:*"
4. Microsoft.DataApiBuilder.exe start