Configured CosmosDB Local Emulator for running integration tests on the Windows agent (#1876)
## Why make this change?

Currently, we are running integration tests against a production
CosmosDB account. Recently, we've encountered 429 errors during these
tests because the same production account is used for each OS agent. To
address this issue, we have started using the CosmosDB Local Emulator
when running integration tests on the Windows agent.

## What is this change?
Changed the Windows agent to windows-latest. This agent has an emulator
preinstalled.
Removed the FileTransform task for the Windows agent. Temporarily
removed the Mac-os job to prevent 429 errors.
Todo: Use an emulator in the Mac-os agent and Linux agent.

## How was this tested?

- [x] Integration Tests

Co-authored-by: Neeraj Sharma (from Dev Box) <neesharma@microsoft.com>