Ensure customers do not have to replicate stored proc parameters in the config. (#1847)
Stored procedure today, is trying to validate that number of params in
config == number of params in sp definition. this makes it difficult for
customer as they have to respecify all the parameters in the sp in the
config again. Our validation should only ever fail in two scenarios:
1. The customer provides inputs in the config that dont match whats
there in the stored proc.
2. The customer does not provide input in the query and there is no
backing default in the config. (ideally if there was a default on the db
stored proc we should have detected it but it seems very hard for sql).
As a workaround, customer can just define that default in the config if
they want.

Difficulties in determining whether a stored procedure parameter has
default values:
1. Documented by microsoft - 
<img width="579" alt="image"
src="https://github.com/Azure/data-api-builder/assets/124841904/7adc3e6d-663f-4694-9321-f7acd3290f56">

https://learn.microsoft.com/en-us/sql/relational-databases/system-catalog-views/sys-sql-modules-transact-sql?view=sql-server-ver16
2. There are some alternatives but they are flaky
-https://copyprogramming.com/howto/get-default-value-of-stored-procedure-parameter

We can try the SQL DOM approach but the focus on this pr is to ensure
that irrespective of whether there is or is not a default parameter
customers can query their endpoint. If there is no query parameter
provided and nothing specified in config we error out (even if for now
they have a default defined in the db.)

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>