Change the form of our NextLink used for pagination cursors (#1851)
## Why make this change?

Updates the form of our NextLink cursor used for pagination. This only
updates the part used for the response, which is then used to continue
the paginated requests. Matches the desired form.

## What is this change?

Rather than using part of the `SqlQueryStructures` to form the cursor,
we create an object that is only used specifically for the NextLink with
respect to the response/request. This means we don't need all of the
fields that are included in the pagination columns, and that the
information contained in this object is really unrelated to the query
structure.

We therefore introduce a new class in our pagination utilities which can
be used to facilitate the serialization and deserialization of the
NextLink cursor in the response/request. Using an anonymous object was
considered, but we re-use this object in enough places, and it holds
complex enough data, that it will be easier to understand and be more
reusable if we define a new class here inside of our pagination
utilities.


The Json structure that is contained within the response/request that we
(de)serialize with this new class is therefore modified.

Previous

```
[{"Value":0,"Direction":0,"TableSchema":"foo","TableName":"bar","ColumnName":"baz"}]")
```

Updated

```
[{"EntityName":"foobar","FieldName":"qux","Value":0,"Direction":0}]
```

Complete NextLink without being decoded

```
"nextLink": "https://localhost:5001/rest/Book?$first=1&$after=W3siVmFsdWUiOjEsIkRpcmVjdGlvbiI6MCwiVGFibGVTY2hlbWEiOiJkYm8iLCJUYWJsZU5hbWUiOiJib29rcyIsIkNvbHVtbk5hbWUiOiJpZCJ9XQ%3D%3D"
```
## How was this tested?

Modified our current test suite to match the new format.

## Sample Request(s)

and paginated request should work as long as it uses or generated a
cursor for pagination.

one such example for when REST has a path of /rest against the Book
entity.

https://localhost:5001/rest/Book?$first=1

---------

Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>