Hot Reload of Runtime Section (#1726)
## Why make this change?

Related to, covering the runtime section of the config only
https://github.com/Azure/data-api-builder/issues/67

Because we are not able to make easy use of the `IOptionsMonitor` which
would have allowed for the updating of individual sections of the config
automatically (ie: just refreshing the `runtime options`), we monitor
and reload the entire file. However, we currently do not have support
for reloading the sections outside of the `Runtime` section of the
config. Because of this, we will not document and announce this feature
until the other sections can be safely reloaded.

Additionally, certain options are used to configure services at startup,
and despite refreshing the `RuntimeConfig`, those services will still be
configured with the values from initialization on startup. This means
adding a mechanism for refreshing those services to make use of the
updated options after a Hot Reload, and that work will come in a future
PR.



## What is this change?

Adds a config file watcher during startup which will monitor for any
file changes to the config and then trigger an update event in the case
that the file is changed.

We instantiate the config file watcher when the runtime provider is
first asked to get a config. We then register a method owned by the
runtime config provider that will be called back to in the case that we
wish to hot reload. However, we only do this for when we start the
service in local development. So that means when we are `not` late
configured, and we are in a Host Mode of Development on startup. In
other cases we do not watch the config file for changes, and do not make
this callback in the case that the config file does change.

To get the location of the file that we are going to watch we use the
config loader's file system and stored data regarding the config file.

## How was this tested?

Unit testing was added to verify correct hot reloading of the runtime
options that we intend to have changing.

## Sample Request(s)

Any valid request after the change ought to work.

---------

Co-authored-by: Sean Leonard <sean.leonard@microsoft.com>