DataWarehouse query support for data-api-builder (#1873)
## Why make this change?
Adds support for Data warehouse to data-api-builder project.

## What is this change?
Created a new DWSQLQueryBuilder. DWSQL reuses metadataprovider,
dbexceptionParser and queryExecutor from mssql as those files would be
same.
Also since we are jsonifying and returning a string, we return the empty
string for null values. For string objects, this means we will return
empty whenever there is null, for all other objects we ensure to
continue to return null.
This change DOES NOT support following:
1. Mutations (Update, delete, execute)
2. Stored procedures

These will come in later prs.

## How was this tested?
1. Manual integration test done with local datawarehouse. 
2. Ran the schema file on a DW to ensure test schema is generated
correctly.
3. Added automated integration tests.
4. Note: With string aggregate a string is returned as the json. The
graphql code however, formats the data according to the return type
expected by the query, hence deep equal to json is not working. For now
the tests are being run against the mssql db and hence the mssql results
can be compared to the return of the dw query executed against the mssql
db.

---------

Co-authored-by: aaronburtle <93220300+aaronburtle@users.noreply.github.com>
Co-authored-by: Aniruddh Munde <anmunde@microsoft.com>