Fix base64 encoding of $after query parameter value in NextLink (#1895)
## Why make this change?

We were not providing a valid base64 encoding of the after cursor in the
`NextLink` for pagination.

Closes https://github.com/Azure/data-api-builder/issues/1799

## What is this change?

Re-factor the logic of creating a `NextLink`. The desired behavior is
for the `after` portion of the `NextLink` to be a valid base 64 encode,
we change the logic to ensure this is the case.

## How was this tested?

Run against our current test suite.

## Sample Request(s)

Any valid pagination request which will have characters that would be
uri encoded to contain invalid base 64 encode characters. Such as,
`https://localhost:5001/rest/Book?$first=1&$orderby=title` which will
include spaces in the `after` portion.