Adds helpful error message for mutations executed without read permissions - SQL database types (#1874)
## Why make this change?

- Closes #1464 
- At the moment, when a mutation operation is executed using a role
which does not have read permissions (and Anonymous does not have read
permissions), the mutation operation succeeds at the database layer but
a generic AuthZ error message is returned. This could lead to the user
concluding that the database operation did not go through because of the
error response.
- This PR adds changes to return a helpful error message to the users -
`The mutation operation {operation name} was successful but the current
user is unauthorized to view the response due to lack of read
permissions`
## What is this change?

- The objective of this change is to return a helpful error message that
lets users know a) what has happened as a result of executing this
mutation request b) the reason as why this error message is returned
- Presence of read permission is evaluated for the current role (with
which the request was executed) as well as Anonymous role (because, for
graphQL, read permissions is inherited from Anonymous role by other
roles). If read permission is inapplicable, then a custom error message
is returned.
 

## How was this tested?
- [x] Integration Testing
- [x] Manual Testing

The request is executed in the context of Anonymous role. Anonymous role
does not have read permissions configured but has the other three -
create, update and delete configured.

**Response:**

![image](https://github.com/Azure/data-api-builder/assets/11196553/fe31bb33-fbe3-472d-8961-4e2b7afaf76f)

**Create operation executed successfully**

![image](https://github.com/Azure/data-api-builder/assets/11196553/bca54e12-9742-4c96-8743-fb93c35def12)


## Error Messages

**Current Format (for mutation without read permission)**

![image](https://github.com/Azure/data-api-builder/assets/11196553/62de9657-b07b-499e-8680-7f37312bc387)


**New Format introduced by this change (for mutation without read
permission)**

![image](https://github.com/Azure/data-api-builder/assets/11196553/93960121-e6f3-4912-a326-aaeb5d78ceb3)

The operation name such as `createbook`/`createPublisher` is included in
the error message to help identify the operation which resulted in error
when there are multiple mutations in the same request

![image](https://github.com/Azure/data-api-builder/assets/11196553/bdfab0fe-dce0-4601-bbc6-be2729cf6335)