Eliminate generation of config files multiple times in the integration test pipelines (#1910)
## Why make this change?

- Closes #1909 
- To include CLI as part of the integration testing, config file used
for all the integration tests are generated by CLI and in turn, these
generated config files are used by the tests. The config file generation
is added as a post build event (in the CLI module) to enable generation
of config files in the a) integration testing pipelines b) local (if
developer wishes to try out the config files we've used during
development/testing.)
- At the moment, the config file generation gets triggered twice.
Ensuring that the config file generation is executed only once will
improve the time taken in the integration testing pipelines.

## What is this change?

- The config file generation gets triggered when building the 1) CLI 2)
CLI.Tests modules.
- The build step in the pipeline is broken up as two

1. Build all the non-testing modules with
`-p:generateConfigFileForDbType=<database_type>`
2. Build the test modules without any arguements. The presence of
`-p:generateConfigFileForDbType` argument determines the execution of
config generation script. So, when the test modules are built without
this argument, the config file generation script does not execute.

## Choice of the executing config generation script in CLI vs CLI.Tests
module:

The PostBuildEvent for executing the config generator scripts after
build could have been added either in `CLI` or `CLI.Tests` module. The
module where this event is added determines where the
`generateConfigFileForDbType` is used along with the `dotnet build`
command.

From the perspective of being able to execute the integration tests,
there is no difference in adding it in CLI or CLI.Tests module.

Adding it in CLI project has the additional advantages: 

1) When a developer wishes to clone the repo and work with the config
files we have used thus far for development/testing, being able to
execute this post build event using the `-p:generateConfigFileForDbType`
enables them to do this. Test module does not seem to be the right place
for this.
2) Being able to validate new changes being worked on manually in local
- For these, typically the test modules are not built. For example: When
a change spans CLI + Engine modules, being able to generate the config
file and validate that the introduced changes work fine is essential.
One should be able to perform this manual validation just by building
the modules where the changes are there without requiring building the
test projects.

## How was this tested?

- [x] Executing integration tests in the pipeline manually. Also, in the
pipelines that have executed for this PR, it can be seen that the config
generation executed only once in the integration testing pipelines.


## Sample Request(s)

N/A